Use one dictionary per file
We had a lot of conflict using a shared `dict` file for all the PRs. We are
using a simpler approach now where we create a dictionary file per .po
translated to avoid this conflicts.

When calling pospell, we concatenate all these files together and use the result
to call pospell properly.