Do not use `make build` on Travis
Travis is giving us a Python executable that it's already inside a virtualenv.
So, using that executable to create another virtualenv does not work properly
and all the packages are installed in the original virtualenv instead in the one
created by the Makefile; which makes packages to not be found at `venv/bin/`.

I'm using Travis `script` option to call the commands that we need just using
the virtualenv they provide to us instead of calling `make build`.