omitir warnings de 'availability' y 'wasm-notavail' (#1847)
the whole story on the 'availability' warnings are that there is a `..
availability::` command in the cpython docs that include, in some cases,
stuff from the 'wasm-availability' file. So we have references "within"
the 'availabiliy' command.

as you can see, the entry has no mention of WASM or wahtsover in some
entries, like "Unix, Windows",
but still will generate the following warning:

```
cpython/Doc/library/os.rst:: WARNING: inconsistent term references in translated message. original: [':ref:`Availability <availability>`', ':ref:`wasm-availability`'], translated: [':ref:`Disponibilidad <availability>`']
```

The reason that due to the internal reference to 'wasm-availability' ,
even in cases like this entry, where it's only Unix and Windows, for
example:

   https://docs.python.org/3.11/library/os.html?highlight=os#os.getenv

it generates the reference to the 'Availability' page, which inside has
the 'wasm-availability' inside:

   https://docs.python.org/3.11/library/intro.html#availability


Additionally, the `wasm-notavail` file is excluded to avoid the warning
related to the file not being inclued in any index. We could have add
`:orphan:` to the file inside `cpython` but we do it via the exclude
patterns to avoid touching files on the submodule.